Update currentResponse on Navigation
Previously, when clicking on a link in a page or submitting a form,
Casper.page would be updated to reflect the newly loaded page, but
Casper.currentResponse would not.  This caused the response passed to
'then' callbacks to only reflect the initial page load.  This patch
resolves that issue so that Casper's page and currentResponse are
consistent and subsequent 'then' call backs receive the response of the
most recent main page load.