Fixes #512: continue to the next step on assertion errors
When an assertion fails, casperjs stops any steps in the current suite.
This could cause issues in a test suite. For example, "withFrame"
generates some steps:

- one to switch to the given frame
- one to execute the given tests
- one to switch to the main frame.

If tests fail, the third step is not executed, and then next tests fail
because they are not executed in the main frame.

we can saw this issue in selftests: if a test fails in tests/suites/casper/frame.js,
tests in tests/suites/casper/global.js fail too.