Refs #482: fixes issue with popup opened by a _blank link
With SlimerJS, checkStep turns into an infinite loop when
clicking on an <a target="_blank"> links, because navigationRequested
is staying always at true. In fact, it should not be modify
to true. In this case, onNavigationRequested is called with
isMainFrame=true instead of false. This is an issue
of SlimerJS, that needs a patch on Gecko.
https://github.com/laurentj/slimerjs/issues/48

This is a "bypass" patch, and should be removed when
SlimerJS will be released with its own patched Gecko.