utils.mergeObjects+gecko: fix copy of properties with getter only
if a property is only a getter, we could have a Javascript error in strict mode when
setting the value to the new object.

"TypeError: setting a property that has only a getter"

We should then define the property with defineProperty, not by setting the value on the
new object.