fix xpath error message
When getElementInfo fails to find an xpath selector,
the following error is reported:

CasperError: Cannot get informations from [object Object]: element not found.

where [object Object] corresponds to the selector. This makes it difficult
to debug; it would be nice if the error message reported the substance
of the xpath selector.