Update installation.rst
If I just run `npm install -g casperjs` I get an error about no compatible targets (full message at http://pastebin.com/zAdpTtdB)

The way to fix this is add the @VERSION to the npm but some people might be too scared of the error message to really understand it. I think the clearer way is to show the possible examples from the doc.

I'd like to include a little information about which version (beta2 or beta3) should be chosen but I don't know enough about the differences.

Cheers!