Added support for multiple file uploads
Added support for multiple file uploads

fixed scope of fileFieldSelector

fixed scope of fileFieldSelector

changed the check for file.path to check if length > 0 instead of 1

When file.path.length was 1 and file.path was an array(when uploading single file to a multiple input) the path would be set as a nested array. The updated behaviour will to do this only if the file.path.length is 0. This should throw a non-existent file error during paths.map

Reverted the matchEngine change

Skip the "file multiple" test on phantomjs 2.0.0

Test fails on phantomjs 2.0.0 due to a fakepath issue

Add tests for command line options

Specifically to catch options incorrectly sent to the engine instead of
the script. See #1384.

Fix assignment of options to engine vs script

Consider entire option name, not just initial chars. Fixes #1384.

fixed scope of fileFieldSelector

Reverted the matchEngine change

Skip the "file multiple" test on phantomjs 2.0.0

Test fails on phantomjs 2.0.0 due to a fakepath issue

Add tests for command line options

Specifically to catch options incorrectly sent to the engine instead of
the script. See #1384.

Fix assignment of options to engine vs script

Consider entire option name, not just initial chars. Fixes #1384.