changed the check for file.path to check if length > 0 instead of 1
When file.path.length was 1 and file.path was an array(when uploading single file to a multiple input) the path would be set as a nested array. The updated behaviour will to do this only if the file.path.length is 0. This should throw a non-existent file error during paths.map