suggestion for PhantomJS API reference in documentation
I was confused by the notation in the documentation here on the PhantomJS API that is available through casperjs.

Before it said `casper.page.casperPageApi()`, which made me think that I had to call that method to get access to the API methods. in reality, those methods are available at `casper.page`. My change makes it a bit clearer for me (I think `.nameOfMethod` suggests that we're calling a method of some-other-name), but I can also see this being written as an example that calls an existing method in PhantomJS's page API.