Modified error message for addListener and prependListener in event.js - refs #1844
Both addListener and prependListener will write an error message if an attempt is made to add a listener that results in the max listener threshold being exceeed. The error message previously referenced node and give an inaccurate recommendation to the user in terms of how to update the maximum number of listeners.

The message has been modified to remove any reference to node and to also indicate how they can update the max listeners using the casper prototype. The message has also been modified to inform the user what the current number of listeners is, the type of listener that it was trying to add and the current maxium number of listeners.